+++ packages/apps/docs/src/components/Layout/components/Header/ThemeToggle.tsx
@@ -28,7 +28,8 @@ export const ThemeToggle: FC = () => {
       onPress={toggleTheme}
       title="Toggle between Light and Dark theme"
       aria-label="Toggle between Light and Dark theme"
-      icon={<MonoContrast />}
-    />
+    >
+      <MonoContrast />
+    </NavHeaderButton>
   );
 };