+++ packages/apps/docs/src/components/Search/components/SearchResults.tsx
@@ -119,7 +119,7 @@ export const SearchResults: FC<IProps> = ({
                   <Stack justifyContent="flex-end">
                     <Link href={`/search?q=${query}`} passHref legacyBehavior>
                       <Button
-                        icon={<MonoChevronRight />}
+                        endVisual={<MonoChevronRight />}
                         title="Go to search results"
                         onPress={state.close}
                       >