+++ packages/apps/graph-client/src/components/graphql-query-dialog/graphql-query-dialog.tsx
@@ -21,7 +21,7 @@ export const GraphQLQueryDialog = (
   return (
     <>
       <Button
-        icon={<MonoInfo />}
+        endVisual={<MonoInfo />}
         title="Show the GraphQL query used."
         isCompact
         variant="transparent"