+++ packages/apps/tools/src/components/Partials/transactions/transfer/sender-details.tsx
@@ -59,7 +59,6 @@ export const SenderDetails: FC<ISenderDetailsProps> = ({
             errorMessage={errors.sender?.message}
             endAddon={
               <Button
-                icon={<MonoContentCopy />}
                 variant="transparent"
                 onPress={async () => {
                   await navigator.clipboard.writeText(field.value);
@@ -67,7 +66,9 @@ export const SenderDetails: FC<ISenderDetailsProps> = ({
                 aria-label="Copy Account Name"
                 title="Copy Account Name"
                 type="button"
-              />
+              >
+                <MonoContentCopy />
+              </Button>
             }
             description={isLedger ? t('unexpected-account-name') : undefined}
           />