+++ packages/tools/kadena-cli/src/prompts/network.ts
@@ -295,8 +295,8 @@ export const networkDeletePrompt: IPrompt<string> = async (
   }
 
   let message = `Are you sure you want to delete the configuration for network "${defaultValue}"?`;
-  if (isNotEmptyString(previousQuestions.isDefaultNetwork)) {
-    message += `\nThis is the "default network". If you delete it, then the default network settings will also be deleted.`;
+  if (previousQuestions.isDefaultNetwork === true) {
+    message += `\nYou have currently set this as your "default network". If you delete it, then the default network settings will also be deleted.`;
   }
 
   return await select({