+++ packages/apps/graph-client/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @kadena/graph-client
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [d82fbd97d]
+  - @kadena/graph@0.1.14
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [fe324038b]
+  - @kadena/graph@0.1.13
+
+## 0.1.15
+
+### Patch Changes
+
+- 6e6b89bdf9: Adjust client to new transaction and signers queries format and
+  also to newly added relational connections on transaction
+- fa2cce1cb: Updated pages and components with new token related objects
+- b3af5a40e: Added Guard to non fungible chain accounts
+- fc1308be85: Removed duplicate queries
+- 1ebfb0a52: Adjusted queries and pages to accomodate new transaction signatures
+  format
+- Updated dependencies [2bac09b5e]
+- Updated dependencies [b316674811]
+- Updated dependencies [8f7b1221c]
+- Updated dependencies [fa60f7fd9]
+- Updated dependencies [6aac4b623]
+- Updated dependencies [89412c156d]
+- Updated dependencies [2a7d0978ee]
+- Updated dependencies [3373d6397e]
+- Updated dependencies [0e5cceacc9]
+- Updated dependencies [e44ac6fdc]
+- Updated dependencies [fa2cce1cb]
+- Updated dependencies [f83bdeac1]
+- Updated dependencies [96c7b98341]
+- Updated dependencies [6514f74a5f]
+- Updated dependencies [7d1dbdd25]
+- Updated dependencies [b3af5a40e]
+- Updated dependencies [1ebfb0a52]
+- Updated dependencies [3b2e0fe7c]
+- Updated dependencies [fc1308be85]
+- Updated dependencies [96e8c2556]
+- Updated dependencies [5a304c326]
+- Updated dependencies [6c2ec7fa9]
+- Updated dependencies [ffb089144]
+- Updated dependencies [fcdcf99f4]
+- Updated dependencies [8bb3168818]
+- Updated dependencies [a605cc1a6]
+- Updated dependencies [b7c5ca5789]
+- Updated dependencies [077c7c1e6]
+- Updated dependencies [5c5937eef6]
+  - @kadena/graph@0.1.12
+  - @kadena/react-ui@0.8.4
+
 ## 0.1.14
 
 ### Patch Changes